Move services to folders
