+++ src/services/graphql/GraphQLService.ts
@@ -11,8 +11,8 @@ import {
   map,
   shareReplay,
 } from 'rxjs/operators';
-import ApiBase, { GraphqlQuery, GraphqlResponse } from '../api-base.js';
-import { ConnectionStatus } from '../model/connection-status.js';
+import ApiBase, { GraphqlQuery, GraphqlResponse } from '../../api-base';
+import { ConnectionStatus } from '../../model/connection-status';
 
 type QueryOrDocument = string | DocumentNode;
